Rewrote i2c driver to support interrupt. Only support byte read and write. Block read and write is not supported any more. This is how it works in interrupt mode: Multiple tasks can call i2c_read_byte_data or i2c_write_byte_data simultaneously. Then they have to compete for a device lock. Only one task will be allowed to proceed to write to device-global buffer. Then it will clear mask of some interrupts and put itself to sleep. Interrupts tick in and get tx request transferred and rx request received. Then mask out all interrupts and wakes the task up. The woken task releases device lock and exits kernel. The other tasks start competing device lock again.

+++ kernel/include/drivers/i2c/galileo_i2c.h
@@ -1,6 +1,6 @@
 #include "kernel.h"
 
-s32 i2c_read_byte_data(u8 command);
+u8 i2c_read_byte_data(u8 command);
 s32 i2c_write_byte_data(u8 command, u8 value);
 s32 i2c_read_block_data(u8 command, u8 length, u8 *values);
 s32 i2c_write_block_data(u8 command, u8 length, u8 *values);
